+++ packages/ref-resolver/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@stoplight/spectral-ref-resolver-v1.0.2](https://github.com/stoplightio/spectral/compare/@stoplight/spectral-ref-resolver-v1.0.1...@stoplight/spectral-ref-resolver-v1.0.2) (2022-10-03)
+
+
+### Bug Fixes
+
+* **ref-resolver:** update @stoplight/json-ref-resolver from 3.1.3 to ~3.1.4 ([dc97f24](https://github.com/stoplightio/spectral/commit/dc97f2414caaaae4c64166122f078c2de9d81ac2))
+
 # [@stoplight/spectral-ref-resolver-v1.0.1](https://github.com/stoplightio/spectral/compare/@stoplight/spectral-ref-resolver-v1.0.0...@stoplight/spectral-ref-resolver-v1.0.1) (2021-09-14)
 
 ### Bug Fixes