+++ packages/rulesets/CHANGELOG.md
@@ -1,3 +1,24 @@
+# [@stoplight/spectral-rulesets-v1.14.1](https://github.com/stoplightio/spectral/compare/@stoplight/spectral-rulesets-v1.14.0...@stoplight/spectral-rulesets-v1.14.1) (2022-10-06)
+
+
+### Bug Fixes
+
+* **rulesets:** handle empty payload and headers in AsyncAPI message's examples validation ([#2284](https://github.com/stoplightio/spectral/issues/2284)) ([4068221](https://github.com/stoplightio/spectral/commit/406822182635dde897ce97fffdc15f23ad35deab))
+
+# [@stoplight/spectral-rulesets-v1.14.0](https://github.com/stoplightio/spectral/compare/@stoplight/spectral-rulesets-v1.13.0...@stoplight/spectral-rulesets-v1.14.0) (2022-10-03)
+
+
+### Features
+
+* **rulesets:** add support for 2.5.0 AsyncAPI ([#2292](https://github.com/stoplightio/spectral/issues/2292)) ([0cb2e85](https://github.com/stoplightio/spectral/commit/0cb2e853147b494b5e6a0c41d2fd891d43e3cf8f))
+
+# [@stoplight/spectral-rulesets-v1.13.0](https://github.com/stoplightio/spectral/compare/@stoplight/spectral-rulesets-v1.12.0...@stoplight/spectral-rulesets-v1.13.0) (2022-09-26)
+
+
+### Features
+
+* **rulesets:** add rule to check if the AsyncAPI document is using the latest version ([#2282](https://github.com/stoplightio/spectral/issues/2282)) ([366779f](https://github.com/stoplightio/spectral/commit/366779f8b3029cbe474df91356ce0ba13ff0f369))
+
 # [@stoplight/spectral-rulesets-v1.12.0](https://github.com/stoplightio/spectral/compare/@stoplight/spectral-rulesets-v1.11.3...@stoplight/spectral-rulesets-v1.12.0) (2022-08-24)